Refactor of code base into modules
- Instead of keeping everything in main, new modules have been added for
  each of the sensors that we'll be using.
- The TCP connections for each of the sensors is now optional. If a connection
  cannot be establish, the thread will just collect sensor data. This is
  to help with testing and for future enhancements where the RC car will
  be doing data processing independent of the data broadcast.
